dockergym/envs/alfworld/requirements.txt
textworld[pddl]>=1.6.1
pyyaml
